Add script to predict the weather this Friday
It uses the Dark Sky API, just like the hubot-forecastio plugin, the
main difference is it's programmed to give you the weather for Friday
instead of the weather right now because for ICFB we want to know if
it's going to rain or not when planning the Friday event.

The API provides hourly data up to 48 hours in advance and from then on
daily data (at some hour, I couldn't see a pattern) up to 7 days in
advance.  This covers all our cases.  If Friday is within 48 hours,
it'll tell you the prediction for 6 PM on Friday (when we usually meet).
If Friday is further in the future it will get the daily prediction for
whatever hour is available on that day.

This script makes heavy usage of the Moment.js library for doing maths
with time, like figuring out when the next Friday is, and which of the
predictions in the API response data match the hour/day we need,
depending on the accuracy.

The regular expression for detecting Friday weather requests is a bit
wobbly, but it's late and I don't care.  It matches:

  Friday weather
  Friday weather in Nehru part
  Friday weather Nehru part
  Friday weather in

The last one is the weird one, but I consider this a feature, not a bug.
When the exact location is omitted, it defaults to Budapest, it's hard
coded, sorry about that.  If I ever publish this script I'll make it
overrideable from the environment.